+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changes
 tip (unreleased)
 ----------------
 
+- Add Hebrew dataset for last names
+- Add Hebrew dataset for first names
 - Modified names.get_first_name to only accept 'male' and 'female' gender values
 
 0.3.0 (2013-05-14)